DestroysEarnedBadges action implemented
