+++ app/services/cancels_course_membership.rb
@@ -1,5 +1,6 @@
 require "light-service"
 require_relative "cancels_course_membership/destroys_assignment_weights"
+require_relative "cancels_course_membership/destroys_earned_badges"
 require_relative "cancels_course_membership/destroys_grades"
 require_relative "cancels_course_membership/destroys_membership"
 require_relative "cancels_course_membership/destroys_submissions"
@@ -50,13 +51,6 @@ class CancelsCourseMembership
     self
   end
 
-  def self.removes_earned_badges(membership)
-    EarnedBadge.for_course(membership.course)
-      .for_student(membership.user)
-      .destroy_all
-    self
-  end
-
   def self.removes_flagged_users(membership)
     FlaggedUser.for_course(membership.course)
       .for_flagged(membership.user)
@@ -66,13 +60,6 @@ class CancelsCourseMembership
 
   def self.removes_group_memberships(membership)
     GroupMembership.for_course(membership.course)
-      .for_student(membership.user)
-      .destroy_all
-    self
-  end
-
-  def self.removes_predicted_earned_badges(membership)
-    PredictedEarnedBadge.for_course(membership.course)
       .for_student(membership.user)
       .destroy_all
     self
