+++ spec/services/cancels_course_membership_spec.rb
@@ -72,22 +72,6 @@ describe CancelsCourseMembership do
     let(:membership) { create(:student_course_membership) }
     let(:student) { membership.user }
 
-    it "removes the earned badges for the student" do
-      another_earned_badge = create :earned_badge, student: student
-      course_earned_badge = create :earned_badge, student: student, course: course
-      described_class.for_student membership
-      expect(student.reload.earned_badges).to eq [another_earned_badge]
-    end
-
-    it "removes the predicted earned badges for the student" do
-      another_earned_badge = create :predicted_earned_badge, student: student
-      course_earned_badge = create :predicted_earned_badge, student: student,
-        badge: create(:badge, course: course)
-      described_class.for_student membership
-      expect(PredictedEarnedBadge.where(student_id: student.id)).to \
-        eq [another_earned_badge]
-    end
-
     it "removes the predicted earned challenges for the student" do
       another_earned_challenge = create :predicted_earned_challenge,
         student: student
